+++ packages/flipper_models/lib/isar/variation_mock.dart
@@ -6,7 +6,6 @@ final variationMock = Variant(
     action: 'create',
     productId: 2,
     unit: 'Per Item',
-    table: 'variants',
     productName: 'Custom Amount',
     branchId: 11,
     supplyPrice: 0.0,
@@ -17,7 +16,6 @@ final variationMock = Variant(
   ..sku = 'sku'
   ..productId = 2
   ..unit = 'Per Item'
-  ..table = 'variants'
   ..productName = 'Custom Amount'
   ..branchId = 11
   ..taxName = 'N/A'